backend/pyproject.toml
[project]
name = "finsight-ai-backend"
version = "0.3.0"
description = "FinSight AI - Financial Intelligence Platform"
requires-python = ">=3.11"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
    "langchain",
    "langchain-google-genai",
    "langchain-community",
    "langchain-chroma",
    "langchain-huggingface",
    "langgraph",
    "chromadb",
    "pypdf",
    "sentence-transformers",
    "httpx",
    "aiofiles",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "sqlalchemy",
    "aiosqlite",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]